chore: ensure apply exits and confirm aborts
