+++ test/typescript/multiple-stacks/cdktf.json
@@ -1,6 +1,6 @@
 {
   "language": "typescript",
-  "app": "npm run --silent compile && node main.js",
+  "app": "npx ts-node main.ts",
   "terraformProviders": [
     "null@ ~> 3.1.0"
   ],
